zcu102-zynqmp.conf: Add variables to provide paths of pmu-firmare deployment
When building pmu firmware using multiconfig setup, the binaries can be
deployed in different build directory. Provide variables to set the
binaries in the required path for QEMU.

Fix SPL dependency on pmu-firmware. Since the pmu-firmware is being built
as a part of multiconfig provide variables to fetch appropriate
pmu-firmware binaries.

Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>